backend/core/views.py
from rest_framework import status, generics, filters
from rest_framework.decorators import api_view, permission_classes
from rest_framework.permissions import IsAuthenticated
from rest_framework.response import Response
from rest_framework.pagination import PageNumberPagination
from django.contrib.auth.models import User
from django.conf import settings
from django.shortcuts import get_object_or_404
from django.core.cache import cache
from django.views.decorators.cache import cache_page
from django.utils.decorators import method_decorator
from django.db.models import Q, Count, Sum, Avg, F
from django.db.models.functions import Coalesce
from django.utils import timezone
from datetime import datetime, timedelta
import logging

from common.decorators import (
    upload_rate_limit, 
    user_rate_limit, 
    cache_response,
    log_performance
)
from common.cache_utils import CacheManager



from content.models import ContentSubmission
from deals.models import Deal
from influencers.models import InfluencerProfile, SocialMediaAccount

logger = logging.getLogger(__name__)






