backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg==3.1.13
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
requests==2.31.0
Pillow==10.4.0
celery==5.3.4
redis==5.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.2.1
django-redis==5.4.0
python-magic==0.4.27
python-magic-bin==0.4.14
psutil==6.1.0#
 Testing dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.1
faker==20.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
coverage==7.3.2